requirements-all.txt
aiofiles==24.1.0
aiosmtplib==3.0.1
aiosqlite==0.20.0
alembic==1.13.3
asyncpg==0.30.0
bleach==6.1.0
croniter>=2.0.0
cryptography==43.0.3
deepgram-sdk==3.7.7
elevenlabs==1.9.0
fastapi==0.115.0
google-cloud-aiplatform>=1.90.0
google-genai>=1.16.0
httpx>=0.28.1
jsonschema>=4.20.0
openai==1.52.0
opentelemetry-exporter-otlp>=1.40.0
opentelemetry-instrumentation-fastapi>=0.49b0
opentelemetry-instrumentation-redis>=0.49b0
opentelemetry-instrumentation-sqlalchemy>=0.49b0
opentelemetry-sdk>=1.40.0
passlib[argon2]==1.7.4
passlib[bcrypt]==1.7.4
pdfplumber==0.11.4
pgvector>=0.3.6
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
prometheus-fastapi-instrumentator==7.0.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]==2.9.2
pyotp==2.9.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-docx==1.1.2
python-dotenv>=1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
qrcode==7.4.2
redis[hiredis]==5.1.1
ruff
sendgrid==6.11.0
sentry-sdk[fastapi]==2.19.2
simpleeval==0.9.13
spacy>=3.7,<4.0
sqlalchemy==2.0.36
stripe>=7.0.0
structlog==24.4.0
tenacity==8.5.0
tiktoken==0.8.0
twilio==9.3.7
uvicorn[standard]==0.31.1
websockets==13.1
